requirements.txt
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
google-api-core==2.19.0
google-api-python-client==2.130.0
google-apps-meet==0.1.6
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
new-py==0.1.2
oauthlib==3.2.2
outcome==1.3.0.post0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyparsing==3.1.2
PySocks==1.7.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0